Module Sc_project.Manager

module Log : Ez_logs.T

val install_include_dir_in : workspace:Sc_core.Types.workspace -> Sc_sys.File.TYPES.dir Lwt.t

val run_info : 'a Types.project -> Types.run

val problem : 'a Types.project -> Types.project_problem

val has_oracle : 'a Types.project -> bool

val seeks_oracle_failures : 'a Types.project -> bool

val entrypoint_name : 'a Types.project -> string

val log_unexpected_errors : (unit -> 'a Lwt.t) -> 'a Lwt.t

val setup_error : Types.setup_error -> 'a

val check_external_tools_availability : unit -> unit Lwt.t

val make_ltest_module : workspace:Sc_core.Types.workspace -> Types.project_config -> (module Sc_ltest.Main.S)

type preprocessed_project = {

workspace : Sc_core.Types.workspace;

config : Types.project_config;

codebase : [ `C ] Sc_sys.File.TYPES.file;

label_data : Sc_ltest.Types.label_data;

given_entrypoint_name : string;

}

val labeling_error : Types.labeling_error -> 'a

val failed_labeling : workdir_status:[ `Created | `Reused ] -> exn -> 'a

val elaboration_error : Types.elaboration_error -> 'a

val failed_elaboration : exn -> 'a

val make_labeling_module : Types.project_config -> (module Sc_ltest.Main.S)

val do_labeling : incdir:Sc_sys.File.dir -> workspace:Sc_core.Types.workspace -> Types.project_config -> preprocessed_project Lwt.t

val skip_labeling : workspace:Sc_core.Types.workspace -> Types.project_config -> preprocessed_project Lwt.t

val setup_project_params : test_repr:'a Sc_corpus.Types.repr -> preprocessed_project -> 'a Types.project_params Lwt.t

val update_label_status : covinfo:Sc_store.Types.covinfo -> 'a Sc_C.Cov_label.t -> 'a Sc_C.Cov_label.t

val update_labels_status : covinfo:Sc_store.Types.covinfo -> Sc_ltest.Types.labels -> Sc_ltest.Types.labels

val write_headers_in : workspace:Sc_core.Types.workspace -> 'a Types.project_params -> ('b Sc_sys__File.TYPES.file * 'c Sc_sys__File.TYPES.file) Lwt.t

val elaborate : test_repr:'a Sc_corpus.Types.repr -> preprocessed_project -> 'a Types.project Lwt.t

val initialize : initialization_options:Sc_core.Types.initialization_options -> test_repr:'a Sc_corpus.Types.repr -> config:Types.project_config -> 'a Types.project Lwt.t

Initializes a project.

An input project is either a raw project to prepare for other tools, or a project already prepared that a tool may want to customize.

val info : 'a Types.project -> (Sc_store.Types.covinfo * Sc_corpus.Types.info) Lwt.t

val report_tool_status : 'a Types.project -> toolname:string -> elapsed_time:float -> tests_generated:int -> (unit, string) Stdlib.Result.t -> unit Lwt.t

val save_tool_statistics : 'a Types.project -> unit

val print_tool_statistics : Stdlib.Format.formatter -> 'a Types.project -> unit