Module Sc_project.Run

val new_in : workdir:Sc_sys.File.dir -> loaded_config_file:Sc_config.Eztoml.file option -> Types.run Lwt.t

new_in ~workdir ~loaded_config_file initializes working directory workdir, dumping relevant runtime configuration along the way. loaded_config_file should be the TOML file (ony one for now) that has been loaded to setup workdir.

val successive_configs : Types.run -> Toml.Types.table Lwt_stream.t

successive_configs run returns a stream of TOML tables with actual run configuration for each successive run before and including run.

val ref_time : Types.run -> float

val ref_time_of : run_num:int -> Types.run -> float