Module Sc_project.Types

Types used to represent and manipulate Seacoral projects.

type project_problem = Sc_ltest.Types.config

A project-specific problem actually corresponds to an Ltest configuration

type project_config = {

project_run : run;
(*
runtime status
*)

project_workspace : Sc_core.Types.workspace;
(*
root workspace
*)

project_check_syntax : bool;
(*
check for syntax errors prior to labeling
*)

project_problem : project_problem;
(*
actual test generation problem
*)

project_pointer_handling : pointer_handling;
(*
parameters for interpreting pointer
*)

project_test_timeout : float;
(*
timeout for each test
*)

project_max_validation_concurrency : int;

project_verbose_validation : bool;

project_inhibit_store_autostop : bool;

project_srcdir_root : Sc_sys.File.TYPES.dir;
(*
root of input source files
*)

}

Configuration given to the project

and pointer_handling = {

treat_pointer_as_array : Sc_C.Types.pointer_refs;

treat_pointer_as_cstring : Sc_C.Types.pointer_refs;

array_size_mapping : Sc_C.Types.pointer_constraints;

}

Pointer-specific configuration

and run = {

run_num : int;
(*
Number of Seacoral runs on the same problem
*)

run_ref_times : float NEL.t;
(*
Reference time of the current run
*)

run_configs : [ `toml ] Sc_sys.File.TYPES.file list;
(*
TOML files with actual configs for previous and current runs (most recent last); |run_configs| = run_num.
*)

}

Info on previous and current runs.

type 'raw_test project_params = {

cil : Cil.file;
(*
CIL representation for the whole C project
*)

cil_typing_info : c_types_info;

typdecls : Sc_values.typdecls;
(*
type declarations, derived from cil
*)

func_repr : Sc_C.Types.func_repr;
(*
representation of the entrypoint function; guaranteed not to be `main`
*)

test_struct : Sc_values.Struct.typ;
(*
struct type that gathers all test inputs for the entrypoing
*)

test_repr : 'raw_test Sc_corpus.Types.repr;
(*
module for test inputs representation
*)

init_func : Sc_C.Types.func_repr option;
(*
optional test initializer
*)

oracle_func : Sc_C.Types.func_repr option;
(*
optional test oracle
*)

seek_oracle_failures : bool;
(*
whether analysis tools should seek tests that fail the oracle
*)

}

Project parameters, computed from the project configuration.

and c_types_info = {

stdbool_needed : bool;
(*
Indicates whether the C code parsed as CIL requires `stdbool.h`
*)

}

and any_project_params =

| A : _ project_params -> any_project_params

type 'raw_test project = {

config : project_config;
(*
given configuration
*)

params : 'raw_test project_params;
(*
computed parameters
*)

workspace : Sc_core.Types.workspace;
(*
(== config.project_workspace)
*)

codebase : [ `C ] Sc_sys.File.TYPES.file;
(*
Never defines or use symbol `main`
*)

types_header : [ `h ] Sc_sys.File.TYPES.file;
(*
C header file with declaration of types from the codebase
*)

func_header : [ `h ] Sc_sys.File.TYPES.file;
(*
C header file with declaration of entrypoint and globals
*)

labels : Sc_ltest.Types.labels;
(*
managed labels (could be in parameters)
*)

label_data : Sc_ltest.Types.label_data;
(*
managed label data
*)

store : Sc_store.t;
(*
store of coverage statuses
*)

decoder : 'raw_test Sc_corpus.Decoder.t;
(*
dynamic raw test decoder
*)

validator : 'raw_test Sc_corpus.Validator.t;

corpus : 'raw_test Sc_corpus.corpus;
(*
corpus of inputs
*)

outdir : Sc_sys.File.TYPES.dir;
(*
output directory for generated C files
*)

covdir : Sc_sys.File.TYPES.dir;
(*
output directory for label-covering C files (subdirectory of outdir)
*)

rtedir : Sc_sys.File.TYPES.dir;
(*
output directory for RTE-triggering C files (subdirectory of outdir)
*)

faildir : Sc_sys.File.TYPES.dir option;
(*
output directory for C files that fail the oracle (subdirectory of outdir)
*)

stats : statistics;
(*
various tool statistics
*)

extra : project_extra;
(*
additional data used for project management
*)

}

A projects' "internal" data, parameterized with the type of stores for handling inputs.

and statistics = Statistics.t

and project_extra = {

given_entrypoint_name : string;
(*
name of the entrypoint originally given (may be "main")
*)

}

type info = Sc_store.Types.covinfo * Sc_corpus.Types.info

Succinct information about the project (coverage status, corpus info).

type tool_stats = Statistics.tool_stats = {

time : float;

status : (unit, string) Stdlib.Result.t;

tests_generated : int;

}

Tool statistics

type testsuite =

| Split_testsuite of {

cov : individual_test_in_testsuite list;

covdir : Sc_sys.File.TYPES.dir;

rte : individual_test_in_testsuite list;

rtedir : Sc_sys.File.TYPES.dir;

fail : individual_test_in_testsuite list option;

faildir : Sc_sys.File.TYPES.dir option;

}

| Combined_testsuite of {

cov : combined_tests_in_testsuite;

covdir : Sc_sys.File.TYPES.dir;

rte : combined_tests_in_testsuite;

rtedir : Sc_sys.File.TYPES.dir;

fail : combined_tests_in_testsuite option;

faildir : Sc_sys.File.TYPES.dir option;

}

and individual_test_in_testsuite = {

file : [ `C ] Sc_sys.File.TYPES.file;

metadata : Sc_corpus.Types.test_metadata;

}

and combined_tests_in_testsuite = {

file : [ `C ] Sc_sys.File.TYPES.file;

metadata : Sc_corpus.Types.test_metadata list;

}

Errors & exceptions

type preproc_error =

| Missing_entrypoint of string

| Missing_cover_target of string

| No_inputs

type setup_error =

| Ltest of Sc_ltest.Types.error

| Preproc of preproc_error NEL.t

exception SETUP_ERROR of setup_error

type labeling_error =

| Failed_labeling of exn Sc_core.Types.generic_preprocessing_error

| Syntax_errors of {

cmd_error : Sc_C.Cmd.error;

stderr_lines : string list;

}

exception LABELING_ERROR of labeling_error

type elaboration_error =

| Failed_elaboration of exn

| Unknown_formals of {

formals : Basics.Strings.t;

func : Sc_C.Types.func_repr;

}

| Unknown_function of {

fun_name : string;

c_file : [ `C ] Sc_sys.File.TYPES.file option;

}

| Unsupported_formals of {

formals : Basics.Strings.t;

func : Sc_C.Types.func_repr;

}

| Unexpected_initialization_function_with_args of {

func : Sc_C.Types.func_repr;

}

exception ELABORATION_ERROR of elaboration_error